Fix vanilla texture "vibrations" for flats
These vibrations are very noticeable at high resolutions if you pay
attention to the floor texture in a corner when moving the mouse slowly.

This fix uses floating point calculations so the texture coordinates are
more exact in the software renderer, as done by PrBoom+.

Note that this will only apply for rendered views taller than 480p, so
it's still possible to use lower resolution and not have these float
operations. After all, the effect is not so noticeable on low resolution